requirements.txt
flask
pandas
openpyxl
python-dotenv
openai
